Moved skeleton validation to mcfg package
