+++ Arma3ObjectBuilder/ui/import_export_mcfg.py
@@ -2,8 +2,8 @@ import bpy
 import bpy_extras
 
 from ..io_mcfg import importer, exporter
+from ..io_mcfg.validator import SkeletonValidator
 from ..utilities import generic as utils
-from ..utilities.validator import Validator
 from ..logger import ProcessLoggerNull
 
 
@@ -106,8 +106,8 @@ class A3OB_OP_export_mcfg(bpy.types.Operator, bpy_extras.io_utils.ExportHelper):
         scene_props = context.scene.a3ob_rigging
         skeleton = scene_props.skeletons[self.skeleton_index]
 
-        validator = Validator(ProcessLoggerNull())
-        if not validator.validate_skeleton(skeleton, False, True):
+        validator = SkeletonValidator(ProcessLoggerNull())
+        if not validator.validate(skeleton, False, True):
             utils.op_report(self, {'ERROR'}, "Invalid skeleton definiton, run skeleton validation for more info")
             return {'FINISHED'}
